refactor(limits): move direction validation
